feat(readme_split): each function description in its own file
This might look loke a horrible amount of changes, but it actually
is not so much at all.

It basically:

- splits each function chapter into its own file in readme_src/*/*.md
- adds a little script that wrapps all of these into README.md

The resulting README.md is allmost the same as it is now.
This commit includes the changes rendering that does to README.md.

The intention is, to have the descriptions a bit more manageable and
avoid mixups in the semantical structure, like I've noticed before.